requirements.txt
fastapi
uvicorn
openai-whisper
torch
requests